Moved errors into their own files.
