+++ src/Core/AdminConsole/Errors/Error.cs
@@ -1,0 +1,3 @@
+﻿namespace Bit.Core.AdminConsole.Errors;
+
+public record Error<T>(string Message, T ErroredValue);
